fix(tjpr): pop aliases deprecados antes de re-passar kwargs no cjsg
cjsg consumia query/termo/data_* via normalize_pesquisa numa copia
local de kwargs e re-passava o **kwargs original para cjsg_download,
que chamava normalize_pesquisa de novo — agora com pesquisa nao-None
e alias ainda em kwargs, levantando ValueError. Aplica o padrao
canonico (normalize_pesquisa + normalize_datas + pop_normalize_aliases
com include_canonical=True) usado em _esaj/base.py:208-211 e
tjsp/client.py:319-324, materializando o dict canonico de datas e
limpando kwargs antes de delegar.

Os testes test_cjsg_{query,termo}_alias_emits_deprecation_warning,
introduzidos no PR #146 como xfail strict apontando o bug, agora
sao testes regulares passando.

Aproveitando o PR, promove urlencoded_body_subset_matcher e
query_param_subset_matcher para tests/_helpers.py: estavam
duplicados byte-a-byte entre test_cjsg_contract.py e
test_cjsg_filters_contract.py do TJPR; reaproveitaveis para os
proximos sub-PRs 1C-c (TJPE/TJRR/TJMG). Tambem remove estado morto
em tests/fixtures/capture/tjpr.py: page_counter era incrementado em
branches onde o valor era descartado.

Suite verde: 813 passed, 19 skipped, 4 xfailed.

Refs #146.
