+++ CHANGELOG.md
@@ -7,9 +7,14 @@ e este projeto adere ao [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+### Fixed
+
+- `TJPRScraper.cjsg` propagava `**kwargs` com aliases deprecados (`query`, `termo`, `data_inicio`, `data_fim`, etc.) para `cjsg_download`, fazendo `normalize_pesquisa` levantar `ValueError` na segunda invocação. Agora pop dos aliases via `pop_normalize_aliases` antes do re-pass — alinhado com `_esaj/base.py` e `tjsp/client.py`. Os testes `test_cjsg_{query,termo}_alias_emits_deprecation_warning` deixam de ser `xfail` e passam diretamente.
+
 ### Added
 
-- Contratos de teste offline para TJPR (`cjsg`), primeira fatia do sub-PR 1C-c. Contrato cobre o fluxo `GET home → POST pesquisa.do (paginação) → GET actionType=exibirTextoCompleto` (chamado uma vez por linha "Leia mais..."), com 3 cenários (`typical_com_paginacao`, `single_page`, `no_results`) + propagação de filtros (`data_julgamento_*`, `data_publicacao_*`) + alias deprecado `data_inicio`/`data_fim`. Aliases `query`/`termo` cobertos como `xfail` documentando bug conhecido em que o `cjsg` re-passa `**kwargs` para `cjsg_download`, fazendo `normalize_pesquisa` levantar `ValueError` na segunda invocação (a corrigir em refactor #84). Samples reais capturados em `tests/tjpr/samples/cjsg/` via `tests/fixtures/capture/tjpr.py`. TJPE/TJRR/TJMG ficam para sub-PRs subsequentes — TJPE adiado por indisponibilidade do backend no momento da captura. Refs #104, #113, #116, #122.
+- Contratos de teste offline para TJPR (`cjsg`), primeira fatia do sub-PR 1C-c. Contrato cobre o fluxo `GET home → POST pesquisa.do (paginação) → GET actionType=exibirTextoCompleto` (chamado uma vez por linha "Leia mais..."), com 3 cenários (`typical_com_paginacao`, `single_page`, `no_results`) + propagação de filtros (`data_julgamento_*`, `data_publicacao_*`) + aliases deprecados `query`/`termo`/`data_inicio`/`data_fim`. Samples reais capturados em `tests/tjpr/samples/cjsg/` via `tests/fixtures/capture/tjpr.py`. TJPE/TJRR/TJMG ficam para sub-PRs subsequentes — TJPE adiado por indisponibilidade do backend no momento da captura. Refs #104, #113, #116, #122.
+- `tests._helpers.urlencoded_body_subset_matcher` e `tests._helpers.query_param_subset_matcher`: matchers do `responses` que validam apenas um subset de campos (body urlencoded e query params, respectivamente), úteis quando o scraper envia tokens dinâmicos (CSRF, ViewState, crypto) que o contrato não deve fixar. Promovidos a partir de duplicação byte-a-byte entre `tests/tjpr/test_cjsg_contract.py` e `tests/tjpr/test_cjsg_filters_contract.py`; ficam disponíveis para os próximos sub-PRs 1C-c (TJPE/TJRR/TJMG).
 - Contratos de teste offline para a familia 1C-a (HTML/Session e JSON simples): TJRN, TJPA, TJRO, TJSC e TJPI (`cjsg`). Contratos via `responses` + samples reais versionados em `tests/<sigla>/samples/cjsg/`, validam schema minimo do DataFrame e payload enviado (`json_params_matcher` para TJRN/TJPA/TJRO, `urlencoded_params_matcher` para TJSC, `query_param_matcher` para TJPI). Scripts de captura em `tests/fixtures/capture/<sigla>.py` importam os novos helpers publicos de `download.py` (regra 12), evitando drift silencioso. TJPA e TJRO aplicam saneamento pos-captura para remover highlights de Elasticsearch e truncar blobs de texto; TJSC cobre a dual-URL (`listar_resultados` em `page=1`, `ajax_paginar_resultado` em `page>1`). Testes de `DeprecationWarning` para aliases `query`/`termo` nos 5 tribunais, `nr_processo` em TJRN/TJRO, e `data_inicio`/`data_fim` em TJPA/TJSC/TJRN/TJRO/TJPI (os cinco tribunais que chamam `normalize_datas` apos o PR #94). Wiring de schema pydantic deliberadamente fora do escopo — seguira junto com o refactor #84 da familia. Refs #94, #104, #113, #116, #120.
 - Helpers publicos `build_cjsg_payload` (TJRN, TJPA, TJRO), `build_cjsg_form_body` + `cjsg_url_for_page` (TJSC) e `build_cjsg_params` (TJPI, agora estendido com `data_min`/`data_max` apos o wiring do PR #94) extraidos de `courts/<sigla>/download.py` e reusados pelos scripts em `tests/fixtures/capture/`. TJPA ganha tambem a constante `CJSG_HEADERS` (antes inline em `_fetch_page`) e o helper publico `post_cjsg(session, payload)` que centraliza URL + serializacao com `ensure_ascii=False` + headers — `_fetch_page` e o capture script chamam o mesmo helper, eliminando o ultimo ponto de drift entre scraper e captura na familia 1C-a.
 - Cobertura de auto-descoberta de paginacao em TJSC e TJPI: novo `test_cjsg_paginas_none_descobre_via_html` exercita `_get_total_pages` (regex `(\d+)\s*documentos?\s*encontrados?` em TJSC; `[?&]page=(\d+)` em TJPI) sobre o sample `single_page`, fechando o gap em que `paginas=1` bypassava o ramo `paginas is None` do `cjsg_download_manager`. TJSC tambem passa a assertir `orgao_julgador` no schema minimo das paginas 1 (a pagina AJAX nao emite o campo, mas a uniao via concat preserva-o no DataFrame final).
